WP CLI: New Module + Regenerate CSS Command
