+++ includes/plugin.php
@@ -519,10 +519,6 @@ class Plugin {
 				new Images_Manager();
 			}
 		}
-
-		if ( defined( 'WP_CLI' ) ) {
-			require_once ELEMENTOR_PATH . '/includes/wp-cli-commands.php';
-		}
 	}
 
 	/**
